backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
transformers==4.30.2
torch==2.2.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
opencv-python==4.8.0.76
pillow==10.0.0
mediapipe==0.10.21
chromadb==0.4.0
sentence-transformers==2.2.2 